requirements.txt
pyyaml
pillow
streamlit
openai
python-docx
python-dotenv
requests
